レリックステータス適用の再設計

+++ NIGHTRAVEL.Shared/Interfaces/Model/Entity/Relic.cs
@@ -19,8 +19,8 @@ namespace NIGHTRAVEL.Shared.Interfaces.Model.Entity
     {
         public int id { get; set; }                     //レリックのID
         public string name { get; set; }                //レリックの名前
-        public double const_effect { get; set; }        //レリックの効果量
-        public double rate_effect { get; set; }         //レリックの効果量
+        public int const_effect { get; set; }        //レリックの効果量
+        public float rate_effect { get; set; }         //レリックの効果量
         public int calculation_method { get; set; }     //計算方法
         public int max {  get; set; }                   //上昇ステータス最大値
         public int status_type {  get; set; }           //上昇ステータスの種類
